% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\name{document_add_attachments}
\alias{document_add_attachments}
\title{Add attachments to existing document}
\usage{
document_add_attachments(doc_id, attachments, api_key = get_api_key())
}
\arguments{
\item{doc_id}{Unique identifier of the document}

\item{attachments}{attachments to attach to the fields in tibble/data.frame
form (one attachment per row), e.g., \code{tibble(field = 7, path = "file.txt")}}

\item{api_key}{RSpace API key}
}
\value{
A JSON object, invisibly. The function will raise an error if \code{doc_id} is not specified.
}
\description{
Add attachments to existing document
}
