% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdtodoc.R
\name{document_create_from_excel}
\alias{document_create_from_excel}
\title{Upload an Excel, csv or tsv document to Rspace}
\usage{
document_create_from_excel(
  path,
  template_id = NULL,
  folder_id = NULL,
  tags = NULL,
  attachment = NULL,
  api_key = get_api_key(),
  existing_document_id = NULL
)
}
\arguments{
\item{path}{Tabular file}

\item{api_key}{RSpace API key}
}
\description{
This function can upload Excel/csv/tabular files to Rspace structured documents.
The file needs to have exactly two columns, one with the Rspace structured document fields and one with the content.
}
\examples{
excel_to_doc_body("assay_with_information.xlsx")
}
