% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdtodoc.R
\name{document_create_from_html}
\alias{document_create_from_html}
\title{Upload a html document to RSpace}
\usage{
document_create_from_html(
  path,
  template_id = NULL,
  folder_id = NULL,
  tags = NULL,
  attachment = NULL,
  existing_document_id = NULL,
  api_key = get_api_key()
)
}
\arguments{
\item{path}{html document to upload}

\item{template_id}{document id of the RSpace template used.
Overwritten by the template of existing_document_id if that one is specified.
A basic document is created if no template is specified.}

\item{folder_id}{Optional. folder_id in which the document will be created (can be a notebook).
The document will be placed in an API inbox if no specific folder is specified.}

\item{tags}{Optional. vector of tags to apply to the document}

\item{attachment}{Optional. attachment to attach to one of the fields, e.g., \code{list(field = 7, path = "file.txt")}}

\item{existing_document_id}{Optional. If you want to replace a document by a new one, specify the current identifier here.}

\item{api_key}{RSpace API key}
}
\value{
Invisible JSON response from the API.
}
\description{
This function can upload a html document (e.g., generated from quarto) to an
RSpace Basic Document, or to a Structured Document if the template is also provided.
}
