% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\name{document_search}
\alias{document_search}
\title{Global search for a term}
\usage{
document_search(query, ..., api_key = get_api_key())
}
\arguments{
\item{query}{description}

\item{...}{query parameters as documented in
\url{https://community.researchspace.com/public/apiDocs} [GET /documents]}

\item{api_key}{RSpace API key}
}
\description{
Global search for a term, works identically to the simple "All" search in RSpace Workspace.
Must be >= 3 characters long.
}
