% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{fields_to_data_frame}
\alias{fields_to_data_frame}
\title{Convert fields list to a tibble}
\usage{
fields_to_data_frame(fields, simplify = T)
}
\arguments{
\item{fields}{fields list as retrieved from RSpace API}

\item{simplify}{Whether to simplify the returned tibble by converting/removing columns}
}
\value{
A tibble with the fields as rows.
}
\description{
Convert fields list to a tibble
}
