% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{file_download}
\alias{file_download}
\title{Download a file from the gallery}
\usage{
file_download(file_id, path = ".", api_key = get_api_key())
}
\arguments{
\item{file_id}{gallery file to be downloaded}

\item{path}{download destination}

\item{api_key}{RSpace API key}
}
\description{
Download a file from the gallery
}
