% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{file_upload}
\alias{file_upload}
\title{Upload a file to the gallery}
\usage{
file_upload(path, api = get_api())
}
\arguments{
\item{path}{file to be uploaded}

\item{api}{RSpace API}
}
\description{
Upload a file to the gallery
}
