% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders.R
\name{folder_tree}
\alias{folder_tree}
\title{List contents of a folder}
\usage{
folder_tree(folder_id = NULL, api_key = get_api_key())
}
\arguments{
\item{folder_id}{Unique identifier of the folder,
if NULL will return contents of the Workspace Home folder}

\item{api_key}{RSpace API key}
}
\description{
List contents of a folder
}
