% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_api}
\alias{get_api}
\title{Get the RSpace API}
\usage{
get_api()
}
\description{
Gets the RSpace API list containing the API key and API URL.
}
\examples{
\dontrun{
get_api()
}

}
\seealso{
\code{\link[=get_api_key]{get_api_key()}}, \code{\link[=get_api_url]{get_api_url()}}
}
