% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_api_key}
\alias{get_api_key}
\title{Get the RSpace API key}
\usage{
get_api_key()
}
\description{
Gets the RSpace API key from the \code{RSPACE_API_KEY} environment variable.
See \code{\link[=set_api_key]{set_api_key()}} on how to set it.
}
\examples{
\donttest{
get_api_key()
}

}
\seealso{
\code{\link[=set_api_key]{set_api_key()}}
}
