% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_api_url}
\alias{get_api_url}
\title{Get the RSpace API URL}
\usage{
get_api_url()
}
\description{
Gets the RSpace API URL from the \code{RSPACE_API_URL} environment variable.
See \code{\link[=set_api_url]{set_api_url()}} on how to set it.
}
\seealso{
\code{\link[=set_api_url]{set_api_url()}}
}
