% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{put_all_fields_in_one_field}
\alias{put_all_fields_in_one_field}
\title{put_all_fields_in_one_field}
\usage{
put_all_fields_in_one_field(doc_body_fields, use_html_sep = T)
}
\arguments{
\item{doc_body_fields}{multiple fields in a list}

\item{use_html_sep}{If \code{TRUE}, each field is placed in a html paragraph, with \if{html}{\out{</p>}} and \if{html}{\out{<p>}}}
}
\value{
a list with one field, with only content, all contents from other fields, separated by \verb{\\n}.
}
\description{
Put a list of fields into one field. This can be needed when no Structured
Document template is specified and a Basic Document is used, but the
html/excel/other input has multiple subheaders or fields.
}
\examples{
\dontrun{
doc_body$fields <- put_all_fields_in_one_field(doc_body$fields)
}

}
\keyword{internal}
