% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{set_api_key}
\alias{set_api_key}
\title{Set the \code{RSPACE_API_KEY} environment variable}
\usage{
set_api_key(rspace_api_key)
}
\arguments{
\item{rspace_api_key}{Your RSpace API key}
}
\description{
Sets the \code{RSPACE_API_KEY} environment variable to the provided RSpace API key.
To create an API key go to 'Manage API Key' section of your profile page (MyRSpace -> Profile).
}
\details{
This will only set the environment variable for the current session,
to set it permanently, add \verb{RSPACE_API_KEY=<your_api_key_here>} to your \code{.Renviron} file,
for example, using \code{\link[usethis:edit]{usethis::edit_r_environ()}}.
}
