% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api_status}
\alias{api_status}
\title{Function to check availability of the API service}
\usage{
api_status(api_key = get_api_key())
}
\arguments{
\item{api_key}{RSpace API key}
}
\description{
Function to check availability of the API service
}
\examples{
\dontrun{
api_status()
}

}
