% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdtodoc.R
\name{document_create_from_excel}
\alias{document_create_from_excel}
\title{Upload a tabular document to RSpace}
\usage{
document_create_from_excel(
  path,
  file_type = NULL,
  document_name = NULL,
  template_id = NULL,
  folder_id = NULL,
  tags = NULL,
  attachments = NULL,
  existing_document_id = NULL,
  api_key = get_api_key()
)
}
\arguments{
\item{path}{tabular file to upload. Can be XLSX, CSV or TSV}

\item{file_type}{an optional character string to specify the file type. Will be guessed from the file name if not specified.}

\item{document_name}{specify the name of the RSpace entry. If not specified,
it will be the value in Title, Name, title, or name if that is one of the fields in the Excel document.
If that does not exist, it will be the file name.}

\item{template_id}{document id of the RSpace template used.
Overwritten by the template of existing_document_id if that one is specified.
A basic document is created if no template is specified.}

\item{folder_id}{Optional. folder_id in which the document will be created (can be a notebook).
The document will be placed in an API inbox if no specific folder is specified.}

\item{tags}{Optional. vector of tags to apply to the document}

\item{attachments}{Optional. attachments to attach to the fields in tibble/data.frame form (one attachment per row), e.g., \code{tibble(field = 7, path = "file.txt")}}

\item{existing_document_id}{Optional. If you want to replace a document by a new one, specify the current identifier here.}

\item{api_key}{RSpace API key}
}
\description{
This function can upload tabular files to RSpace structured documents.
The file needs to have exactly two columns, one with the RSpace structured document fields and one with the content.
}
