% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdtodoc.R
\name{document_create_from_html}
\alias{document_create_from_html}
\title{Upload a html document to Rspace}
\usage{
document_create_from_html(
  path,
  template_id = NULL,
  folder_id = NULL,
  tags = NULL,
  attachment = NULL,
  api_key = get_api_key(),
  existing_document_id = NULL
)
}
\arguments{
\item{path}{html document to upload}

\item{template_id}{document id of the RSpace template used. Overwritten by the template of existing_document_id if that one is specified.}

\item{folder_id}{folder_id in which the document will be created (can be a notebook)}

\item{tags}{vector of tags to apply to the document}

\item{attachment}{attachment to attach to one of the fields, e.g., \code{list(field = 7, path = "file.txt")}}

\item{api_key}{RSpace API key}

\item{existing_document_id}{if you want to replace a document by a new one, specify the current identifier here.}
}
\description{
This function can upload a html document (e.g., generated from quarto) to an
Rspace Basic Document, or to a Structured Document if the template is also provided.
}
