% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\name{document_list_attachments}
\alias{document_list_attachments}
\title{List attachments of a structured document}
\usage{
document_list_attachments(doc_id, api_key = get_api_key())
}
\arguments{
\item{doc_id}{Unique identifier of the document}

\item{api_key}{RSpace API key}
}
\value{
description A tibble with identifiers and information on attachments, one attachment per row.
Returns an empty tibble if no files are attached to the field.
}
\description{
This function lists all attachments of a field in a structured document.
}
