% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\name{document_retrieve}
\alias{document_retrieve}
\title{Get document from RSpace based on document ID}
\usage{
document_retrieve(doc_id, api_key = get_api_key())
}
\arguments{
\item{doc_id}{Unique identifier of the document}

\item{api_key}{RSpace API key}
}
\description{
Get document from RSpace based on document ID
}
