% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{file_upload}
\alias{file_upload}
\title{Upload a file to the gallery}
\usage{
file_upload(path, api_key = get_api_key())
}
\arguments{
\item{path}{file to be uploaded}

\item{api_key}{RSpace API key}
}
\description{
Upload a file to the gallery
}
