% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{put_all_fields_in_one_field}
\alias{put_all_fields_in_one_field}
\title{Put a list of all fields into one field.}
\usage{
put_all_fields_in_one_field(doc_body_fields)
}
\arguments{
\item{doc_body_fields}{multiple fields in a list}
}
\value{
a list with one field containing all content from all fields
}
\description{
This can be needed when no Structured Document template is specified and a
Basic Document is used, but the html/excel/other input has multiple
subheaders or fields.
}
\keyword{internal}
